Change parse method depended on file type

+++ parser/src/main/java/com/scriptorium/censorship/parser/boundary/BookListLoader.java
@@ -1,6 +1,8 @@
 package com.scriptorium.censorship.parser.boundary;
 
 import com.scriptorium.censorship.common.model.BookDto;
+import com.scriptorium.censorship.common.model.ContentParams;
+import com.scriptorium.censorship.common.model.ContentParams.ContentParamsBuilder;
 import com.scriptorium.censorship.parser.util.BookXlsMapper;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
@@ -10,6 +12,7 @@ import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 
@@ -27,7 +30,7 @@ public final class BookListLoader {
     private BookListLoader() {
     }
 
-    public static long getFileSize(String urlAddress) {
+    public static ContentParams loadContentParams(String urlAddress) {
         try {
             URL urlObject = new URL(urlAddress);
             HttpURLConnection urlConnection = (HttpURLConnection) urlObject.openConnection();
@@ -35,23 +38,38 @@ public final class BookListLoader {
             int status = urlConnection.getResponseCode();
             if (status != HttpURLConnection.HTTP_OK) {
                 LOG.error("Response code is: {}", status);
-                return 0;
+                return null;
             }
-            return urlConnection.getContentLengthLong();
+
+            ContentParamsBuilder contentParamsBuilder = ContentParams.builder();
+            String contentName = urlConnection.getHeaderField("Content-Disposition");
+            if (contentName.endsWith("xlsx")) {
+                LOG.debug("Source file is 'xlsx'");
+                contentParamsBuilder.xlsx(true);
+            } else if (contentName.endsWith("xls")) {
+                contentParamsBuilder.xlsx(false);
+                LOG.debug("Source file is 'xls'");
+            } else {
+                LOG.error("Who knows what the extension of this file!");
+                return null;
+            }
+            contentParamsBuilder.lastModified(LocalDateTime.now())
+                    .fileSize(urlConnection.getContentLengthLong());
+            return contentParamsBuilder.build();
         } catch (MalformedURLException e) {
             LOG.error("The URL is not correct: {}", urlAddress);
-            return 0;
+            return null;
         } catch (IOException e) {
             LOG.error("Error establish connection: {}", e.getMessage());
-            return 0;
+            return null;
         }
     }
 
-    public static List<BookDto> loadFileFromUrl(String urlAddress) {
-        return loadFileFromUrl(urlAddress, ROWS_TO_OMIT);
+    public static List<BookDto> loadDataFromUrl(String urlAddress, boolean expectXlsx) {
+        return loadDataFromUrl(urlAddress, ROWS_TO_OMIT, expectXlsx);
     }
 
-    public static List<BookDto> loadFileFromUrl(String urlAddress, int rowsToOmit) {
+    public static List<BookDto> loadDataFromUrl(String urlAddress, int rowsToOmit, boolean expectXlsx) {
         LOG.debug("Trying to get file from URL = {}", urlAddress);
         URL urlObject;
         try {
@@ -61,20 +79,28 @@ public final class BookListLoader {
             return Collections.emptyList();
         }
 
-        File tempFile;
+        File tempFile = null;
         try {
             tempFile = File.createTempFile("loadedGoods-", "-tmp");
             FileUtils.copyURLToFile(urlObject, tempFile, 5000, 120000);
             LOG.info("Successfully save into file {}", tempFile.getName());
             List<BookDto> bookDtoList;
-            bookDtoList = BookXlsMapper.parseXlsFile(tempFile, rowsToOmit);
-            if (!tempFile.delete()) {
-                LOG.error("Cannot delete temporary file!");
+            if (expectXlsx) {
+                bookDtoList = BookXlsMapper.parseXlsxFile(tempFile, rowsToOmit);
+            } else {
+                bookDtoList = BookXlsMapper.parseXlsFile(tempFile, rowsToOmit);
             }
             return bookDtoList;
         } catch (IOException e) {
             LOG.error("Error loading or parsing file: {}", e.getMessage());
             return Collections.emptyList();
+        } finally {
+            if (tempFile != null) {
+                if (!tempFile.delete()) {
+                    LOG.error("Cannot delete temporary file!");
+                }
+            }
         }
+
     }
 }
